+++ src/main/java/courgette/api/CourgetteTestOutput.java
@@ -1,0 +1,7 @@
+package courgette.api;
+
+public enum CourgetteTestOutput {
+    DISCARD,
+    CONSOLE,
+    FILE
+}