+++ src/main/java/courgette/runtime/CourgettePickleMatcher.java
@@ -3,7 +3,6 @@ package courgette.runtime;
 import io.cucumber.core.filter.Filters;
 import io.cucumber.core.gherkin.Feature;
 import io.cucumber.core.gherkin.Pickle;
-import io.cucumber.gherkin.Location;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -32,8 +31,8 @@ public class CourgettePickleMatcher {
         return matched.get();
     }
 
-    public Location matchLocation(int pickleLocationLine) {
-        final Location[] location = {null};
+    public CucumberPickleLocation matchLocation(int pickleLocationLine) {
+        final CucumberPickleLocation[] location = {null};
 
         List<Pickle> pickles = feature.getPickles();
 
@@ -42,7 +41,7 @@ public class CourgettePickleMatcher {
                     .findFirst()
                     .ifPresent(pickleEvent -> {
                         if (filters.test(pickleEvent)) {
-                            location[0] = new Location(pickleEvent.getLocation().getLine(), pickleEvent.getLocation().getColumn());
+                            location[0] = new CucumberPickleLocation(pickleEvent.getLocation().getLine(), pickleEvent.getLocation().getColumn());
                             throw new ConditionSatisfiedException();
                         }
                     });