+++ src/main/java/courgette/runtime/CourgetteTestStatistics.java
@@ -63,6 +63,22 @@ public class CourgetteTestStatistics {
         return failed > 0;
     }
 
+    protected void printToConsole(CourgetteProperties courgetteProperties) {
+        System.out.println("\n───────────────────────────────────────────────────");
+        System.out.println("\u001B[32m" + "             Courgette Test Statistics          " + "\u001B[0m");
+        System.out.println("───────────────────────────────────────────────────");
+        System.out.println("Summary:\t" + passedPercentage() + "% passed" + (failed > 0 ? ", " + failedPercentage() + "% failed" : ""));
+        System.out.println("Duration:\t" + duration);
+        System.out.println("Run Level:\t" + (courgetteProperties.isFeatureRunLevel() ? "Feature" : "Scenario"));
+        System.out.println("Total:\t\t" + total);
+        System.out.println("Passed:\t\t" + passed);
+        System.out.println("Failed:\t\t" + failed);
+        if (courgetteProperties.getCourgetteOptions().rerunFailedScenarios()) {
+            System.out.println("Rerun:\t\t" + rerun);
+        }
+        System.out.println("───────────────────────────────────────────────────\n");
+    }
+
     private void calculateTestStatistics(List<CourgetteRunResult> runResults, CourgetteProperties courgetteProperties) {
         total = (int) runResults.stream().filter(result -> !result.getStatus().equals(CourgetteRunResult.Status.RERUN)).count();
 
@@ -82,7 +98,7 @@ public class CourgetteTestStatistics {
                 TimeUnit.MILLISECONDS.toSeconds(elapsedMill) - TimeUnit.MINUTES.toSeconds(TimeUnit.MILLISECONDS.toMinutes(elapsedMill)));
     }
 
-    private int calculateStatus(List<CourgetteRunResult> runResults, CourgetteRunResult.Status ...status) {
+    private int calculateStatus(List<CourgetteRunResult> runResults, CourgetteRunResult.Status... status) {
         return (int) runResults.stream().filter(result -> Arrays.asList(status).contains(result.getStatus())).count();
     }