+++ src/main/java/courgette/runtime/CucumberPickleLocation.java
@@ -1,0 +1,20 @@
+package courgette.runtime;
+
+public class CucumberPickleLocation {
+
+    private final int line;
+    private final int column;
+
+    public CucumberPickleLocation(int line, int column) {
+        this.line = line;
+        this.column = column;
+    }
+
+    public int getLine() {
+        return line;
+    }
+
+    public int getColumn() {
+        return column;
+    }
+}