+++ src/main/java/courgette/runtime/utils/JacksonUtils.java
@@ -1,0 +1,26 @@
+package courgette.runtime.utils;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.cfg.ConstructorDetector;
+import com.fasterxml.jackson.databind.json.JsonMapper;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
+
+public final class JacksonUtils {
+    public static ObjectMapper CUCUMBER_OBJECT_MAPPER = JsonMapper.builder()
+            .addModule(new Jdk8Module())
+            .addModule(new ParameterNamesModule(JsonCreator.Mode.PROPERTIES))
+            .constructorDetector(ConstructorDetector.USE_PROPERTIES_BASED)
+            .serializationInclusion(JsonInclude.Include.NON_ABSENT)
+            .enable(SerializationFeature.WRITE_ENUMS_USING_TO_STRING)
+            .enable(DeserializationFeature.READ_ENUMS_USING_TO_STRING)
+            .enable(DeserializationFeature.USE_LONG_FOR_INTS)
+            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+            .disable(JsonGenerator.Feature.AUTO_CLOSE_TARGET)
+            .build();
+}